+++ src/autostore/database.py
@@ -1,10 +1,14 @@
 """Database connection."""
 
 from pathlib import Path
+from typing import TypeVar
 
-from sqlmodel import Session, SQLModel, create_engine
+from sqlalchemy.exc import SQLAlchemyError
+from sqlmodel import Session, SQLModel, create_engine, select
 
 from .models import *  # noqa: F403
+
+ModelT = TypeVar("ModelT", bound=SQLModel)
 
 
 class Database:
@@ -38,6 +42,80 @@ class Database:
     def session(self) -> Session:
         """Create a new database session."""
         return Session(self.engine)
+
+    def write(self, *, row: type[ModelT]) -> int:
+        """
+        Write row to database.
+
+        Parameters
+        ----------
+        model
+            Instance of a database model class.
+
+        Returns
+        -------
+        row_id
+            id corresponding to entry in model table.
+        """
+        try:
+            with self.session() as session:
+                session.add(row)
+                session.commit()
+                session.refresh(row)
+                return row.id  # ty:ignore[unresolved-attribute]
+
+        except SQLAlchemyError as e:
+            msg = f"Failed to write {row = } to database."
+            raise RuntimeError(msg) from e
+
+    def fetch(
+        self, *, model: type[ModelT], row_id: int
+    ) -> ModelT | type[ModelT] | None:
+        """
+        Fetch rows based on row id.
+
+        Parameters
+        ----------
+        model
+            Database model class, e.g. CalculationRow or GeometryRow.
+        row_id
+            id corresponding to entry in model table.
+
+        Returns
+        -------
+        model
+            Instance of a database model class.
+        """
+        with self.session() as session:
+            return session.get(model, row_id)
+
+    def query(
+        self, *, model: type[ModelT], **attributes: float | str | None
+    ) -> list[int | None]:
+        """
+        Query for existing rows based on Class attributes.
+
+        Parameters
+        ----------
+        model
+            Database model class, e.g. CalculationRow or GeometryRow.
+        **attributes
+            Database model class attributes, e.g. id = 1 or energy = -0.568.
+
+        Returns
+        -------
+        row_ids
+            ids corresponding to entries in model table.
+        """
+        with self.session() as session:
+            statement = select(model)
+
+            # Append Select constructs to statement
+            for key, value in attributes.items():
+                if hasattr(model, key):
+                    statement = statement.where(getattr(model, key) == value)
+
+            return [getattr(row, "id", None) for row in session.exec(statement).all()]
 
     def close(self) -> None:
         """Close the database connection.