+++ tests/test_autostore.py
@@ -1,13 +1,14 @@
 """autostore tests."""
 
+import json
 from collections.abc import Iterator
+from pathlib import Path
 
-import numpy as np
 import pytest
 from automol import Geometry
-from qcio import Results
+from qcio import CalcType, Results
 
-from autostore import Calculation, Database, fetch, qc, write
+from autostore import Calculation, Database, models, qc
 
 
 @pytest.fixture
@@ -30,9 +31,18 @@ def water() -> Geometry:
 
 
 @pytest.fixture
+def h2() -> Geometry:
+    """Water geometry fixture."""
+    return Geometry(
+        symbols=["H", "H"],
+        coordinates=[[0, 0, 0], [0, 0.74, 0]],  # ty:ignore[invalid-argument-type]
+    )
+
+
+@pytest.fixture
 def xtb_calculation() -> Calculation:
     """XTB calculation fixture."""
-    return Calculation(program="crest", method="gfn2")
+    return Calculation(program="crest", method="gfnff")
 
 
 @pytest.fixture
@@ -61,16 +71,74 @@ def water_xtb_energy_results() -> Results:
     )
 
 
+@pytest.fixture
+def h2_gfnff_stationary_results() -> Results:
+    """Water energy calculation results fixture."""
+    with (Path(__file__).parent / "stationary.json").open(encoding="utf-8") as f:
+        data = json.load(f)
+
+    return Results.model_validate(data)
+
+
 def test_energy(
     water: Geometry,
-    xtb_calculation: Calculation,
     water_xtb_energy_results: Results,
     database: Database,
 ) -> None:
-    """Test writing and reading of the energy."""
-    final_energy = water_xtb_energy_results.data.energy
-    calc_row, geo_row = qc.results.rows(water_xtb_energy_results)
-    write.energy(final_energy, calc_row=calc_row, geo_row=geo_row, db=database)
-    energy = fetch.energy(water, xtb_calculation, hash_name="minimal", db=database)
-    assert energy is not None
-    assert np.isclose(energy.value, -5.062316802835694), f"{energy = }"
+    """Test writing and reading of the energy and corresponding database rows."""
+    # Instantiate GeometryRow, write to database, and ensure correct hash population
+    geom_row = models.GeometryRow(**water.model_dump())
+    geom_id = database.write(row=geom_row)  # ty:ignore[invalid-argument-type]
+    assert database.query(model=models.GeometryRow, hash=water.hash)[0] == geom_id
+
+    # Instantiate CalculationRow, set input geometry id, write to database,
+    # and ensure correct calctype
+    calc_row = qc.results.calc_row(water_xtb_energy_results)
+    calc_row.input_geometry_id = geom_id
+    calc_id = database.write(row=calc_row)  # ty:ignore[invalid-argument-type]
+    assert (
+        database.query(model=models.CalculationRow, calctype=CalcType.energy)[0]
+        == calc_id
+    )
+
+    # Instantiate EnergyRow, set geometry and calculation ids, write to database,
+    # and ensure correct energy value
+    ene_row = models.EnergyRow(
+        geometry_id=geom_id,
+        calculation_id=calc_id,
+        value=water_xtb_energy_results.data.energy,
+    )
+    ene_id = database.write(row=ene_row)  # ty:ignore[invalid-argument-type]
+    assert database.query(model=models.EnergyRow, value=-5.062316802835694)[0] == ene_id
+
+
+def test_stationary(
+    h2: Geometry,
+    h2_gfnff_stationary_results: Results,
+    database: Database,
+) -> None:
+    """Test writing and reading of the energy and corresponding database rows."""
+    # Instantiate GeometryRow, write to database, and ensure correct hash population
+    input_geom_row = models.GeometryRow(**h2.model_dump())
+    input_geom_id = database.write(row=input_geom_row)  # ty:ignore[invalid-argument-type]
+    assert database.query(model=models.GeometryRow, hash=h2.hash)[0] == input_geom_id
+
+    # Instantiate CalculationRow, set input geometry id, write to database,
+    # and ensure correct calctype
+    calc_row = qc.results.calc_row(h2_gfnff_stationary_results)
+    calc_row.input_geometry_id = input_geom_id
+    calc_id = database.write(row=calc_row)  # ty:ignore[invalid-argument-type]
+    assert (
+        database.query(model=models.CalculationRow, calctype=CalcType.optimization)[0]
+        == calc_id
+    )
+
+    # Instantiate EnergyRow, set geometry and calculation ids, write to database,
+    # and ensure it's equal to input model
+    stp_row = models.StationaryPointRow(
+        geometry_id=input_geom_id, calculation_id=calc_id, order=1
+    )
+    stp_id = database.write(row=stp_row)  # ty:ignore[invalid-argument-type]
+    stp_row_fetch = database.fetch(model=models.StationaryPointRow, row_id=stp_id)
+
+    assert stp_row == stp_row_fetch