+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
+### Added
+- CalculationInputGeometryLink table to ensure that CalculationRow objects are linked to input GeometryRow objects.
 
 ## [0.0.4] - 2026-04-01